fix(brush): expose DabRenderer stroke flags and fix an Int-to-Float
Third compile fix round from the on-device build of 0d3c0c5:

- DabRenderer squareStroke/noAntialias were private but assigned from
  EngineRenderer's dabRenderer.apply blocks (STAMP and WET paths) -> 4 errors.
  Make both public like the neighbouring edgeDarkening/wetness flags.
- expandWetDirtyBounds passed the Int WET_WASH_MARGIN_PX to
  Float.coerceAtLeast (Kotlin has no implicit widening) -> 1 error; use
  WET_WASH_MARGIN_PX.toFloat().

Also audited the same error class (cross-class private access and Int->Float
coerces in WetSimMath/WetSimulationRenderer) and found no further cases.

+++ app/src/main/java/com/wetinknext/engine/core/EngineRenderer.kt
@@ -1559,7 +1559,7 @@ class EngineRenderer(
         )
         val margin = (WET_WASH_MARGIN_PX + WET_WASH_MARGIN_SCALE * spread * diag)
             .coerceAtMost(0.12f * diag)
-            .coerceAtLeast(WET_WASH_MARGIN_PX)
+            .coerceAtLeast(WET_WASH_MARGIN_PX.toFloat())
         val m = margin.toInt()
         dirtyBounds[0] = (dirtyBounds[0] - m).coerceAtLeast(0)
         dirtyBounds[1] = (dirtyBounds[1] - m).coerceAtLeast(0)
